requirements.txt
Django==1.11.6
djangorestframework==3.7.1
psycopg2==2.7.3.1
pytz==2017.2
django-crontab==0.7.1
vk==2.0.2
Pillow==4.3.0
pymorphy2==0.8
feedparser==5.2.1
pymorphy2-dicts-ru
python-dateutil
pymongo
mongoengine
flask-mongoengine
